+++ mimosa/cpu-foreach.hh
@@ -10,6 +10,6 @@ namespace mimosa
    *
    * Thread created in cb will by default keep the same cpu affinity.
    */
-  void cpuForeach(const std::function<void ()>& cb, bool affinity = false, int ratio = 1);
+  void cpuForeach(const std::function<void ()>& cb, bool affinity = false, size_t ratio = 1);
 }
